pyproject.toml
[tool.poetry]
name = "manager-dependencies"
version = "0.1.0"
description = ""
authors = ["Otavio <otavio.dantas@syngenta.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
